RPS source code
The source code of the rps game. I encapsulated the game in a class. I
used private variables to hold moves, choices, and options. My public
methods  enabled smooth gameplay especially once I put them in the run
method. The asci art was not perfect but it was a nice touch.